+++ src/commands/uncommentAllLogMessages.ts
@@ -1,20 +1,12 @@
 import * as vscode from 'vscode';
-import { DebugMessage } from '../debug-message';
-import { Command, ExtensionProperties, Message } from '../entities';
+import { Command, Message } from '../entities';
 
 export function uncommentAllLogMessagesCommand(): Command {
   return {
     name: 'turboConsoleLog.uncommentAllLogMessages',
-    handler: async (
-      {
-        delimiterInsideMessage,
-        logMessagePrefix,
-        logType,
-        logFunction,
-      }: ExtensionProperties,
-      jsDebugMessage: DebugMessage,
-      args?: unknown[],
-    ) => {
+    handler: async ({ extensionProperties, debugMessage, args }) => {
+      const { logFunction, logType, logMessagePrefix, delimiterInsideMessage } =
+        extensionProperties;
       const editor: vscode.TextEditor | undefined =
         vscode.window.activeTextEditor;
       if (!editor) {
@@ -22,7 +14,7 @@ export function uncommentAllLogMessagesCommand(): Command {
       }
       const document: vscode.TextDocument = editor.document;
 
-      const logMessages: Message[] = jsDebugMessage.detectAll(
+      const logMessages: Message[] = debugMessage.detectAll(
         document,
         logFunction,
         logType,