+++ src/pro/index.ts
@@ -1,0 +1,2 @@
+import { TurboProFreemiumTreeProvider } from './TurboProFreemiumTreeProvider';
+export { TurboProFreemiumTreeProvider };